+++ src/main.rs
@@ -41,7 +41,7 @@ mod packfile;
 
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord)]
 enum LayerType {
-    Libc(Architecture),
+    Dependency(Architecture, String),
     Java(Architecture),
     Modloader,
     Download(String),
@@ -53,8 +53,9 @@ enum LayerType {
 impl LayerType {
     fn applicable_for(&self, arch: Architecture) -> bool {
         match self {
-            LayerType::Libc(a) => arch == *a,
+            LayerType::Dependency(a, _) => arch == *a,
             LayerType::Java(a) => arch == *a,
+            // For layer types that don't have an architecture, they're always to be included
             _ => true,
         }
     }
@@ -289,7 +290,11 @@ async fn main() -> anyhow::Result<()> {
     // is okay, vs using ARCs, because the size of the cloned things are very small.
 
     for arch in &arches {
-        join_set.spawn(install_musl_layer(*arch, oci_blob_dir.clone()));
+        join_set.spawn(deb::install_debian_package(oci_blob_dir.clone(), deb::Package {
+            name: "libc6".to_string(),
+            version: "2.36-9+deb12u8".to_string(),
+            arch: *arch,
+        }));
         join_set.spawn(install_jre(
             *arch,
             oci_blob_dir.clone(),
@@ -984,219 +989,6 @@ async fn install_jre<P: AsRef<Path>, S: AsRef<str>>(
     Ok(BuiltLayer {
         blob: Some(jre_layer),
         layer_type: LayerType::Java(arch),
-        extra_config: None,
-    })
-}
-
-async fn install_musl_layer<P: AsRef<Path>>(
-    arch: Architecture,
-    oci_blob_dir: P,
-) -> anyhow::Result<BuiltLayer> {
-    ////////////////////////////////
-    //// MUSL
-    ////////////////////////////////
-    // TODO verify the checksum of this DEB file
-    let musl_dep = deb::Package {
-        name: "musl".to_string(),
-        version: "1.2.5-1.1".to_string(),
-        arch,
-    };
-    info!(
-        deb_version = musl_dep.version,
-        url = musl_dep
-            .url()
-            .map(|u| u.as_str().to_string())
-            .unwrap_or("<error>".to_string()),
-        "Installing musl"
-    );
-    let musl_deb_request = reqwest::get(musl_dep.url()?).await?;
-    let mut musl_tar_stream = tokio_util::io::StreamReader::new(
-        musl_deb_request
-            .bytes_stream()
-            .map_err(|e| std::io::Error::new(ErrorKind::Other, e)),
-    );
-    let mut musl_layer_builder = TarLayerBuilder::new(&oci_blob_dir).await?;
-    musl_layer_builder
-        .append_directory(
-            "/lib",
-            &FileInfo {
-                mode: 0o755,
-                uid: 0,
-                gid: 0,
-                last_modified: 0,
-            },
-        )
-        .await
-        .context("Failed to append directory for /lib")?;
-    // TODO This is the worst code I've ever written
-    // throw away the first 120 bytes
-    debug!("Skipping 120 bytes of file header, debian-binary file, and control.tar.xz header");
-    let mut file_header: [u8; 120] = [0; 120];
-    (&mut musl_tar_stream)
-        .read_exact(&mut file_header)
-        .await
-        .context("Reading DEB file header from stream")?;
-
-    let mut control_file_size_bytes: [u8; 10] = [0; 10];
-    (&mut musl_tar_stream)
-        .read_exact(&mut control_file_size_bytes)
-        .await
-        .context("Reading control file size")?;
-    let control_file_size_string = std::str::from_utf8(&control_file_size_bytes)
-        .context("Parse control file size as string")?;
-    debug!(s = control_file_size_string, "string");
-    let control_file_size = control_file_size_string
-        .split(" ")
-        .next()
-        .ok_or(anyhow::anyhow!("No size in ar header"))?
-        .parse::<u64>()
-        .context("Parse control file size string as int")?;
-    debug!(
-        bytes = hex::encode_upper(control_file_size_bytes),
-        size = control_file_size,
-        "Control file size found"
-    );
-    // Another 6 bytes to finish the header
-    debug!("skipping final 2 bytes of control.tar.xz header");
-    tokio::io::copy(&mut (&mut musl_tar_stream).take(2), &mut tokio::io::sink()).await?;
-    // Start of the control.tar.xz file data
-    debug!(
-        size = control_file_size,
-        "skipping control.tar.xz file data bytes"
-    );
-    tokio::io::copy(
-        &mut (&mut musl_tar_stream).take(control_file_size),
-        &mut tokio::io::sink(),
-    )
-    .await?;
-    debug!("skipping 48 bytes of data.tar.zx header");
-    // skip through the start of the data.tar.zx file header to the size
-    tokio::io::copy(&mut (&mut musl_tar_stream).take(48), &mut tokio::io::sink()).await?;
-    let mut data_file_size_bytes: [u8; 10] = [0; 10];
-    (&mut musl_tar_stream)
-        .read_exact(&mut data_file_size_bytes)
-        .await
-        .context("Reading data file size")?;
-    debug!(
-        bytes = hex::encode_upper(data_file_size_bytes),
-        "read data file bytes"
-    );
-    let data_file_size_string =
-        std::str::from_utf8(&data_file_size_bytes).context("Parse data file bytes as string")?;
-    let data_file_size = data_file_size_string
-        .split(" ")
-        .next()
-        .ok_or(anyhow::anyhow!("No size in ar header"))?
-        .parse::<u64>()
-        .context("Parse data file length from string to int")?;
-    debug!(
-        bytes = hex::encode_upper(data_file_size_bytes),
-        size = data_file_size,
-        "data file size found"
-    );
-    // Another 6 bytes to finish the header
-    debug!("Skipping final 2 bytes of data.tar.zx header");
-    tokio::io::copy(&mut (&mut musl_tar_stream).take(2), &mut tokio::io::sink()).await?;
-    debug!("Opening xz data stream");
-    // data.tar.xz data stream
-    let compressed_musl_data_stream = &mut (&mut musl_tar_stream).take(data_file_size);
-    let mut buffered_compressed_musl_data_stream = BufReader::new(compressed_musl_data_stream);
-    let mut musl_data_stream =
-        BufReader::new(XzDecoder::new(&mut buffered_compressed_musl_data_stream));
-    loop {
-        let mut header_bytes: [u8; 512] = [0; 512];
-        (&mut musl_data_stream)
-            .read_exact(&mut header_bytes)
-            .await?;
-        if header_bytes == [0; 512] {
-            debug!("Read 512 nil bytes, end of stream");
-            break;
-        }
-        let header = tar::Header::from_byte_slice(&header_bytes);
-        let size = header.size().context("Header had invalid size")?;
-        let path = header.path().context("Header had invalid path")?;
-        let entry_type = header.entry_type();
-        debug!(
-            size = size,
-            path = path.as_os_str().to_str(),
-            "Parsed data.tar.xz TAR header"
-        );
-        if entry_type == EntryType::Regular
-            && (path.ends_with("libc.so") || path.starts_with(Path::new("./usr/share/doc/")))
-        {
-            if path.ends_with("libc.so") {
-                debug!(
-                            path = ?path,
-                            "Found shared library");
-                musl_layer_builder
-                    .append_file(
-                        format!("/lib/ld-musl-{}.so.1", arch.linux()),
-                        &FileInfo {
-                            mode: 0o0755,
-                            uid: 0,
-                            gid: 0,
-                            last_modified: 0,
-                        },
-                        size,
-                        &mut (&mut musl_data_stream).take(size),
-                    )
-                    .await
-                    .context("Failed to append library file for MUSL")?;
-            } else {
-                debug!(
-                            path = ?path,
-                            "Found documentation");
-                let mut container_path = Path::new("/").to_path_buf();
-                container_path.push(path.to_path_buf().to_path_buf());
-                musl_layer_builder
-                    .append_file(
-                        &container_path,
-                        &FileInfo {
-                            mode: 0o0644,
-                            uid: 0,
-                            gid: 0,
-                            last_modified: 0,
-                        },
-                        size,
-                        &mut (&mut musl_data_stream).take(size),
-                    )
-                    .await
-                    .context("Failed to append documentation/license file for MUSL")?;
-            }
-            let remaining = 512 - (size % 512) as usize;
-            if remaining < 512 {
-                let mut padding_bytes: [u8; 512] = [0; 512];
-                (&mut musl_data_stream)
-                    .read_exact(&mut padding_bytes[..remaining])
-                    .await?;
-                if padding_bytes != [0; 512] {
-                    warn!(
-                        padding_bytes = remaining,
-                        read_bytes = hex::encode_upper(&padding_bytes[..remaining]),
-                        "Padding bytes were not nul bytes, maybe corrupt file?"
-                    )
-                }
-            }
-        } else {
-            debug!("Skipping data.tar.xz entry {:?}", path);
-            let expected_blocks = (size / 512) + (if size % 512 == 0 { 0 } else { 1 });
-            for _ in 0..expected_blocks {
-                (&mut musl_data_stream)
-                    .read_exact(&mut header_bytes)
-                    .await?;
-            }
-        }
-    }
-
-    let musl_layer = musl_layer_builder.finalise().await?;
-    info!(
-        path = ?musl_layer.path,
-        digest = musl_layer.digest(),
-        "Created MUSL Layer"
-    );
-    Ok(BuiltLayer {
-        blob: Some(musl_layer),
-        layer_type: LayerType::Libc(arch),
         extra_config: None,
     })
 }
